feat: cleanup interrupts and traps

+++ src/memlayout.rs
@@ -20,12 +20,12 @@
 use crate::riscv::{MAXVA, PGSIZE};
 
 // qemu UART registers
-pub const UART0: u32 = 0x1000_0000;
-pub const UART0_IRQ: u32 = 10;
+pub const UART0: usize = 0x1000_0000;
+pub const UART0_IRQ: usize = 10;
 
 // virtio mmio interface
-pub const VIRTIO0: u32 = 0x1000_1000;
-pub const VIRTIO0_IRQ: u32 = 1;
+pub const VIRTIO0: usize = 0x1000_1000;
+pub const VIRTIO0_IRQ: usize = 1;
 
 // Core Local Interrupter (CLINT), contains the timer
 pub const CLINT: usize = 0x200_0000;
@@ -37,17 +37,17 @@ pub const fn clint_mtimecmp(hart: usize) -> usize {
 
 // Platform-Level Interrupt Controller (PLIC)
 // https://github.com/riscv/riscv-plic-spec/blob/master/riscv-plic.adoc
-pub const PLIC: u32 = 0x0C00_0000;
-pub const PLIC_PRIORITY: u32 = (PLIC + 0x0);
-pub const PLIC_PENDING: u32 = (PLIC + 0x1000);
+pub const PLIC: usize = 0x0C00_0000;
+pub const PLIC_PRIORITY: u32 = (PLIC + 0x0) as u32;
+pub const PLIC_PENDING: u32 = (PLIC + 0x1000) as u32;
 pub const fn PLIC_SENABLE(hart: usize) -> u32 {
-    PLIC + 0x2080 + (hart as u32 * 0x100)
+    (PLIC + 0x2080 + (hart * 0x100)) as u32
 }
 pub const fn PLIC_SPRIORITY(hart: usize) -> u32 {
-    PLIC + 0x201000 + (hart as u32 * 0x2000)
+    (PLIC + 0x201000 + (hart * 0x2000)) as u32
 }
 pub const fn PLIC_SCLAIM(hart: usize) -> u32 {
-    PLIC + 0x201004 + (hart as u32 * 0x2000)
+    (PLIC + 0x201004 + (hart * 0x2000)) as u32
 }
 
 // Physical RAM resides from KERNBASE to PHYSTOP
